fix(app): update buttons layout in protocol setup page
update buttons layout in protocol setup page

close AUTH-2464
