+++ app/src/organisms/Desktop/Devices/ProtocolRun/ProtocolRunHeader/RunHeaderContent/LabeledValue/index.tsx
@@ -1,12 +1,11 @@
 import {
   COLORS,
-  DIRECTION_COLUMN,
-  Flex,
-  SPACING,
   StyledText,
 } from '@opentrons/components'
 
 import type { ReactNode } from 'react'
+
+import styles from './labeledvalue.module.css'
 
 interface LabeledValueProps {
   label: string
@@ -15,8 +14,8 @@ interface LabeledValueProps {
 
 export function LabeledValue(props: LabeledValueProps): JSX.Element {
   return (
-    <Flex flexDirection={DIRECTION_COLUMN} gridGap={SPACING.spacing4}>
-      <StyledText as="h6" color={COLORS.grey60}>
+    <div className={styles.container}>
+      <StyledText desktopStyle='bodyDefaultRegular' color={COLORS.grey60}>
         {props.label}
       </StyledText>
       {typeof props.value === 'string' ? (
@@ -24,6 +23,6 @@ export function LabeledValue(props: LabeledValueProps): JSX.Element {
       ) : (
         props.value
       )}
-    </Flex>
+    </div>
   )
 }
